requirements.txt
fastapi
uvicorn
boto3 
langchain
langchain-community
langchain-aws
pinecone-client>=3.0.0
langchain-pinecone
python-dotenv
